+++ src/main/java/net/frozenblock/wilderwild/datagen/tag/WWBiomeTagProvider.java
@@ -1177,6 +1177,17 @@ public final class WWBiomeTagProvider extends FrozenBiomeTagProvider {
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_CRIMSON_SHELF_FUNGI_RARE)
 			.add(Biomes.WARPED_FOREST);
 
+		this.getOrCreateTagBuilder(WWBiomeTags.HAS_FALLEN_CRIMSON_FUNGI)
+			.add(Biomes.CRIMSON_FOREST);
+
+		this.getOrCreateTagBuilder(WWBiomeTags.HAS_SNAPPED_CRIMSON_FUNGI)
+			.add(Biomes.CRIMSON_FOREST);
+
+		this.getOrCreateTagBuilder(WWBiomeTags.HAS_FALLEN_WARPED_FUNGI)
+			.add(Biomes.WARPED_FOREST);
+
+		this.getOrCreateTagBuilder(WWBiomeTags.HAS_SNAPPED_WARPED_FUNGI)
+			.add(Biomes.WARPED_FOREST);
 
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_RAINFOREST_MUSHROOM)
 			.addOptional(WWWorldgen.RAINFOREST)